+++ src/pint/models/timing_model.py
@@ -552,6 +552,26 @@ class TimingModel:
                 raise ValueError(f"Unknown kind {kind!r}")
         return c
 
+    def get_params_of_component_type(self, component_type):
+        """Get a list of parameters belonging to a component type.
+
+        Parameters
+        ----------
+        component_type : "PhaseComponent", "DelayComponent", "NoiseComponent"
+
+        Returns
+        -------
+        list
+        """
+        component_type_list_str = "{}_list".format(component_type)
+        if hasattr(self, component_type_list_str):
+            component_type_list = getattr(self, component_type_list_str)
+            return [
+                param for component in component_type_list for param in component.params
+            ]
+        else:
+            return []
+
     def set_param_values(self, fitp):
         """Set the model parameters to the value contained in the input dict.
 
@@ -1762,9 +1782,20 @@ class TimingModel:
         We decide to add minus sign here in the design matrix, so the fitter
         keeps the conventional way.
         """
+
+        noise_params = self.get_params_of_component_type("NoiseComponent")
+        # unfrozen_noise_params = [
+        #     param for param in noise_params if not getattr(self, param).frozen
+        # ]
+
+        # The entries for any unfrozen noise parameters will not be
+        # included in the design matrix as they are not well-defined.
+
         params = ["Offset"] if incoffset else []
         params += [
-            par for par in self.params if incfrozen or not getattr(self, par).frozen
+            par
+            for par in self.params
+            if (incfrozen or not getattr(self, par).frozen) and par not in noise_params
         ]
 
         F0 = self.F0.quantity  # 1/sec